seperate controls and dancer group controls almost done
